MRELEASE-273: a NPE can hide another
git-svn-id: https://svn.apache.org/repos/asf/maven/release/trunk@721313 13f79535-47bb-0310-9956-ffa450edef68

+++ maven-release-manager/src/main/java/org/apache/maven/shared/release/phase/AbstractRewritePomsPhase.java
@@ -785,11 +785,14 @@ public abstract class AbstractRewritePomsPhase
                 MavenProject project = (MavenProject) i.next();
 
                 File pomFile = ReleaseUtil.getStandardPom( project );
-                File file =
-                    new File( pomFile.getParentFile(), pomFile.getName() + "." + pomSuffix );
-                if ( file.exists() )
+                // MRELEASE-273 : if no pom
+                if ( pomFile != null )
                 {
-                    file.delete();
+                    File file = new File( pomFile.getParentFile(), pomFile.getName() + "." + pomSuffix );
+                    if ( file.exists() )
+                    {
+                        file.delete();
+                    }
                 }
             }
         }
